% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_aou_connect.R
\name{mock_aou_disconnect}
\alias{mock_aou_disconnect}
\title{Disconnect from the mock database and clear the stored connection}
\usage{
mock_aou_disconnect(con = getOption("aou.default.con"), shutdown = TRUE)
}
\arguments{
\item{con}{The connection to close. Defaults to \code{getOption("aou.default.con")}.}

\item{shutdown}{Whether to shut down the DuckDB instance. Default \code{TRUE}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Disconnect from the mock database and clear the stored connection
}
