% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{sql_translation.mock_aou_connection}
\alias{sql_translation.mock_aou_connection}
\title{dbplyr SQL translation for mock All of Us (DuckDB) connections}
\usage{
\method{sql_translation}{mock_aou_connection}(con)
}
\arguments{
\item{con}{A \code{mock_aou_connection}.}
}
\value{
A \code{dbplyr} \code{sql_variant}.
}
\description{
Extends DuckDB's dbplyr translation with BigQuery-compatible functions
(currently \code{date_diff()}/\code{DATE_DIFF()}) used by All of Us analysis code.
Not called directly.
}
\keyword{internal}
